Добавляет конфигурацию для создания бинов игры (GameConfiguration), Добавляет слой Repository. Изменяет entity и скрипты sql. Добавляет Мапперы для получения бина Action / SuddenAction / Card / Combination, зная имя (из БД)

+++ backend/server/src/main/resources/db/changelog/game/process/table/schema.creation-1.0.sql
@@ -13,7 +13,7 @@ CREATE TABLE player
 --changeset barabanov:card_deck
 CREATE TABLE card_deck
 (
-    id              SERIAL        PRIMARY KEY,
+    id              BIGSERIAL        PRIMARY KEY,
     value           VARCHAR []
 );
 
@@ -21,7 +21,7 @@ CREATE TABLE card_deck
 --changeset barabanov:card_reset
 CREATE TABLE card_reset
 (
-    id              SERIAL        PRIMARY KEY,
+    id              BIGSERIAL        PRIMARY KEY,
     value           VARCHAR []
 );
 
@@ -31,16 +31,16 @@ CREATE TABLE game_state
 (
     id                 BIGSERIAL              PRIMARY KEY,
     step_quantity      INT,
-    now_turn_id        BIGINT                               REFERENCES player (id),
-    card_deck_id       INT                                  REFERENCES card_deck (id) ON DELETE RESTRICT,
-    card_reset_id      INT                                  REFERENCES card_reset (id) ON DELETE RESTRICT
+    now_turn_id        BIGINT                                  REFERENCES player (id),
+    card_deck_id       BIGINT                                  REFERENCES card_deck (id) ON DELETE RESTRICT,
+    card_reset_id      BIGINT                                  REFERENCES card_reset (id) ON DELETE RESTRICT
 );
 
 
 --changeset barabanov:player_queue_pointer
 CREATE TABLE player_queue_pointer
 (
-    id                       SERIAL        PRIMARY KEY,
+    id                       BIGSERIAL        PRIMARY KEY,
     game_state_id            BIGINT                            REFERENCES game_state (id)  ON DELETE CASCADE,
     pointing_player_id       BIGINT                            REFERENCES player (id),
     pointed_at_player_id     BIGINT                            REFERENCES player (id),
